Animation Tool 제작, Map Editor 구조 개선, Title Level 구현

+++ GameClient/CStage01.cpp
@@ -7,7 +7,7 @@ CStage01::CStage01()
 {
 	CTexture* pTex = CAssetMgr::GetInst()->LoadTexture(L"VOLCANO_01", L"texture\\Volcano_01.png");
 	SetImg(pTex);
-	SetScale(pTex->GetWidth(), pTex->GetHeight());
+	SetScale((float)pTex->GetWidth(), (float)pTex->GetHeight());
 }
 
 CStage01::~CStage01()
